src/backend/requirements.txt
alembic
aniso8601
asgiref
asttokens
attrs
bcrypt
cachetools
certifi
cffi
charset-normalizer
click
colorama
coverage
cryptography
executing
fastapi
google-api-core
google-auth
googleapis-common-protos
graphene
graphql-core
graphql-relay
greenlet
h11
icecream
idna
iniconfig
mako
markupsafe
opencensus
opencensus.ext.azure
opencensus-context
packaging
passlib
pip
pluggy
protobuf
psutil
psycopg2-binary
py
pyasn1
pyasn1-modules
pycparser
pydantic
pygments
PyJWT
pyparsing
pytest
pytest-cov
pytest-mock
python-dateutil
python-dotenv
python-editor
pytz
requests
rsa
setuptools
six
SQLAlchemy
sqlalchemy2-stubs
sqlmodel
starlette
starlette-graphene3
toml
tomli
typing-extensions
urllib3